feat(business): improve i18n support and refactor receipt form layout
- Remove required field asterisks from business form labels
- Add data-i18n attributes for internationalization support on verify button
- Refactor receipt creation form to use dynamic template-based item inputs
- Replace static product/price/vat fields with reusable receipt item template
- Format chart period button attributes for better readability

+++ js/dashboard-notifications.js
@@ -32,7 +32,7 @@ function formatDate(dateStr) {
     const lang = getNotifLang();
     const locale = lang === 'ru' ? 'ru-RU' : 'en-US';
     const date = new Date(dateStr);
-    return date.toLocaleDateString(locale, { day: 'numeric', month: 'short', year: 'numeric' });
+    return date.toLocaleDateString(locale, { day: 'numeric', month: 'numeric', year: 'numeric' });
 }
 
 async function loadNotifications(userId, client) {
